Enum pdb::Error

pub enum Error {
    UnrecognizedFileFormat,
    InvalidPageSize(u32),
    PageReferenceOutOfRange(u32),
    StreamNotFound(u32),
    IoError(Error),
    UnexpectedEof,
    UnimplementedFeature(&'static str),
    SymbolTooShort,
    UnimplementedSymbolKind(u16),
    InvalidTypeInformationHeader(&'static str),
    TypeTooShort,
    TypeNotFound(TypeIndex),
    TypeNotIndexed(TypeIndex, TypeIndex),
    UnimplementedTypeKind(u16),
    UnexpectedNumericPrefix(u16),
}

An error that occurred while reading or parsing the PDB.

Variants
UnrecognizedFileFormat

The input data was not recognized as a MSF (PDB) file.

InvalidPageSize(u32)

The MSF header specifies an invalid page size.

PageReferenceOutOfRange(u32)

MSF referred to page number out of range.

This likely indicates file corruption.

StreamNotFound(u32)IoError(Error)

An IO error occurred while reading from the data source.

UnexpectedEof

Unexpectedly reached end of input.

UnimplementedFeature(&'static str)

This data might be understandable, but the code needed to understand it hasn't been written.

SymbolTooShort

A symbol record's length value was impossibly small.

UnimplementedSymbolKind(u16)

Support for symbols of this kind is not implemented.

InvalidTypeInformationHeader(&'static str)

The type information header was invalid.

TypeTooShort

A type record's length value was impossibly small.

TypeNotFound(TypeIndex)

Type not found.

TypeNotIndexed(TypeIndex, TypeIndex)

Type not indexed -- the requested type (.0) is larger than the maximum TypeIndex covered by the TypeFinder (.1).

UnimplementedTypeKind(u16)

Support for types of this kind is not implemented.

UnexpectedNumericPrefix(u16)

Variable-length numeric parsing encountered an unexpected prefix.