Struct pdb::PDBInformation

pub struct PDBInformation<'s> {
    pub version: HeaderVersion,
    pub signature: u32,
    pub age: u32,
    pub guid: Uuid,
    pub names_offset: usize,
    pub names_size: usize,
    // some fields omitted
}

A PDB info stream header parsed from a stream.

The PDB information stream contains the GUID and age fields that can be used to verify that a PDB file matches a specific binary, as well a list of named PDB streams with their stream indices.

Fields
version: HeaderVersion

The version of the PDB format in use.

signature: u32

A 32-bit timestamp.

age: u32

The number of times this PDB file has been written.

guid: Uuid

A Uuid generated when this PDB file was created that should uniquely identify it.

names_offset: usize

The offset of the start of the stream name data within the stream.

names_size: usize

The size of the stream name data, in bytes.

Methods

impl<'s> PDBInformation<'s>

pub fn stream_names(&self) -> Result<StreamNames>

Get a StreamNames object that can be used to iterate over named streams contained within the PDB file.

This can be used to look up certain PDB streams by name.

Example

let file = std::fs::File::open("fixtures/self/foo.pdb")?;
let mut pdb = pdb::PDB::open(file)?;
let info = pdb.pdb_information()?;
let names = info.stream_names()?;
let mut v: Vec<_> = names.iter().map(|n| n.name.to_string()).collect();
v.sort();
assert_eq!(&v, &["mystream", "/LinkInfo", "/names", "/src/headerblock"]);