Struct pdb::SymbolIter

pub struct SymbolIter<'t> { /* fields omitted */ }

A SymbolIter iterates over a SymbolTable, producing Symbols.

Symbol tables are represented internally as a series of records, each of which have a length, a type, and a type-specific field layout. Iteration performance is therefore similar to a linked list.

Methods

impl<'t> SymbolIter<'t>

pub fn new(buf: ParseBuffer<'t>) -> SymbolIter