Struct pdb::TypeProperties

pub struct TypeProperties(_);

Methods

impl TypeProperties

pub fn packed(&self) -> bool

Indicates if a type is packed via #pragma pack or similar.

pub fn constructors(&self) -> bool

Indicates if a type has constructors or destructors.

pub fn overloaded_operators(&self) -> bool

Indicates if a type has any overloaded operators.

pub fn is_nested_type(&self) -> bool

Indicates if a type is a nested type, e.g. a union defined inside a class.

pub fn contains_nested_types(&self) -> bool

Indicates if a type contains nested types.

pub fn overloaded_assignment(&self) -> bool

Indicates if a class has overloaded the assignment operator.

pub fn overloaded_casting(&self) -> bool

pub fn forward_reference(&self) -> bool

Indicates if a type is a forward reference, i.e. an incomplete Type that serves as a placeholder until a complete Type can be built. This is necessary for e.g. self-referential data structures, but other more common declaration/definition idioms can cause forward references too.

pub fn scoped_definition(&self) -> bool

pub fn has_unique_name(&self) -> bool

pub fn sealed(&self) -> bool

pub fn hfa(&self) -> u8

pub fn intrinsic_type(&self) -> bool

pub fn mocom(&self) -> u8